Addition of instruction functions and adjustment of existing functions:
- Implement functions in compare_and_jump and f_arithmetic_logic.
- Implement the function int get_immediate(CPU *cpu) which reads 64 bits at a given address in the binary file.
- Adjust the incrementation of the cpu->IP variable during the execution of an instruction.

Remaining tasks:
- Implement the 'mov' function that moves bits from a scalar register to another scalar register (same too for vector register).
- Add cases where CF[1] can be set to 1 (an unsigned subtraction with the first source operand lower than the second source operand).
- Improve handling of forbidden memory access errors in the load and store files.
- Introduce benchmarks to measure the performance of opcode functions.
- Enrich unit tests to cover a wider range of scenarios for opcode functions.

+++ vm/operation/src/load.c
@@ -12,8 +12,7 @@
  */
 void loadu(CPU *cpu, Instruction inst)
 {
-    // TODO : overflow
-    u8 offset = get_offset(inst.unused);
+    u16 offset = get_offset(inst.unused);
     u64 addr = cpu->U[inst.source_1] + cpu->U[inst.source_2] + offset;
     if (addr >= MEMORY_SIZE)
     {
@@ -33,7 +32,7 @@ void loadu(CPU *cpu, Instruction inst)
  */
 void loads(CPU *cpu, Instruction inst)
 {
-    u8 offset = get_offset(inst.unused);
+    u16 offset = get_offset(inst.unused);
     u64 addr = cpu->U[inst.source_1] + cpu->U[inst.source_2] + offset;
 
     if (addr >= MEMORY_SIZE)
@@ -54,7 +53,7 @@ void loads(CPU *cpu, Instruction inst)
  */
 void loadf(CPU *cpu, Instruction inst)
 {
-    u8 offset = get_offset(inst.unused);
+    u16 offset = get_offset(inst.unused);
     u64 addr = cpu->U[inst.source_1] + cpu->U[inst.source_2] + offset;
 
     if (addr >= MEMORY_SIZE)
@@ -75,7 +74,7 @@ void loadf(CPU *cpu, Instruction inst)
  */
 void loadv(CPU *cpu, Instruction inst)
 {
-    u8 offset = get_offset(inst.unused);
+    u16 offset = get_offset(inst.unused);
     u64 addr = cpu->U[inst.source_1] + cpu->U[inst.source_2] + offset;
 
     if (addr >= MEMORY_SIZE)
@@ -99,7 +98,7 @@ void loadv(CPU *cpu, Instruction inst)
  */
 void loadt(CPU *cpu, Instruction inst)
 {
-    u8 offset = get_offset(inst.unused);
+    u16 offset = get_offset(inst.unused);
     u64 addr = cpu->U[inst.source_1] + cpu->U[inst.source_2] + offset;
 
     if (addr >= MEMORY_SIZE)
@@ -123,7 +122,7 @@ void loadt(CPU *cpu, Instruction inst)
  */
 void loadg(CPU *cpu, Instruction inst)
 {
-    u8 offset = get_offset(inst.unused);
+    u16 offset = get_offset(inst.unused);
     u64 addr = cpu->U[inst.source_1] + cpu->U[inst.source_2] + offset;
 
     if (addr >= MEMORY_SIZE)
